backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
pymupdf
numpy
redis
python-dotenv
pydantic
huggingface_hub